pyproject.toml
[project]
name = "spark-backend"
version = "0.1.0"
description = "Spark — real-time, context-aware local commerce offers"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "httpx>=0.28",
    "numpy>=2.0",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "neo4j>=5.20",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pyright>=1.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pip-audit>=2.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
